requirements.txt
pandas
numpy
tensorflow
scikit-learn
joblib
requests
python-dotenv
matplotlib
